code commit@rahul
works on session management

+++ src/main/java/com/hrms/controller/LeaveApprovalController.java
@@ -56,8 +56,9 @@ public class LeaveApprovalController {
 	@GetMapping("/leaveApproval")
 	public String leaveApproval(Model model, HttpSession session) {
 
-		System.out.println("leave approval methods");
-		
+		if(session.getAttribute("username")==null) {
+			return "redirect:" + "./";
+		}
 		
 		List<Department> listDepartment = departmentService.getAllDepartments();
 		List<LeaveRequest> listLeaveApproval = leaveRequestService.getEmployeeByStatusN();
@@ -82,6 +83,8 @@ public class LeaveApprovalController {
 	@GetMapping("/leaverequest/{deptCode}")
 	public List<LeaveRequest> getAllLeaveRequestBydept(@PathVariable("deptCode") String deptCode) {
 		 
+		
+		
 		return leaveRequestService.findAllByDeptCodeAndStatusN(deptCode);
 		
 	}
@@ -91,6 +94,11 @@ public class LeaveApprovalController {
 	@GetMapping("/approveLeaveRequest/{leaveRequestId}/{status}")
 	public String approveLeaveRequest(@PathVariable("leaveRequestId") String leaveid, @PathVariable("status") String status,
 			Model model ,HttpSession session) {
+		
+		if(session.getAttribute("username")==null) {
+			return "redirect:" + "./";
+		}
+		
 		session.setAttribute("imgUtil", new ImageUtil());
 		String userCode = (String) session.getAttribute("username");
 		LeaveRequest leaveRequest = leaveRequestService.findLeaveRequestById(Long.valueOf(leaveid));
@@ -133,6 +141,10 @@ public class LeaveApprovalController {
 	@GetMapping(value = { "/deleteLeaveApproval/{id}" })
 	public String deleteActivity(@PathVariable("id") Long id, Model model, HttpSession session) {
 		
+		if(session.getAttribute("username")==null) {
+			return "redirect:" + "./";
+		}
+		
 		String userCode = (String) session.getAttribute("username");
 		this.leaveRequestService.removeLeaveRequest(id);
 		session.setAttribute("username", userCode);
@@ -144,6 +156,9 @@ public class LeaveApprovalController {
 	public String viewLeaveRequestByEmpId(@PathVariable("id")String leaveRequestId,
 						Model model,HttpSession session) {
 		
+		if(session.getAttribute("username")==null) {
+			return "redirect:" + "./";
+		}
 		
 		LeaveRequest leaveRequest = this.leaveRequestService.findLeaveRequestById(Long.parseLong(leaveRequestId));
 		
@@ -158,12 +173,5 @@ public class LeaveApprovalController {
 		model.addAttribute("myhref", "leaveApproval");
 		return "viewLeaveRequest";
 	}
-	
-	
-	
-	
-	
-	
-	
 
 }
